+++ src/includes/functions.php
@@ -59,15 +59,15 @@ function get_avarage_payout($db_link)
     $db_query = 'SELECT * FROM logs';
     $db_result = mysqli_query($db_link, $db_query);
 
-    $payout_average = 0;
+    $payout_average = (float)0;
 
     while ($row = mysqli_fetch_array($db_result)) {
-        $payout_average = $payout_average + $row['payout'];
+        $payout_average = $payout_average + (float)$row['payout'];
     }
 
     $am = mysqli_num_rows($db_result);
 
-    return $payout_average / $am;
+    return (float)($payout_average / $am);  
 }
 
 /**
@@ -79,13 +79,13 @@ function get_daily_payout($db_link)
     $db_query = 'SELECT * FROM logs WHERE DATE(date) = DATE(NOW())';
     $db_result = mysqli_query($db_link, $db_query);
 
-    $payout_daily = 0;
+    $payout_daily = (float)0;
 
     while ($row = mysqli_fetch_array($db_result)) {
-        $payout_daily = $payout_daily + $row['payout'];
+        $payout_daily = $payout_daily + (float)$row['payout'];
     }
 
-    return $payout_daily;
+    return (float)$payout_daily;
 }
 
 /**
@@ -99,11 +99,11 @@ function get_total_payout($db_link)
     $db_query = 'SELECT * FROM logs';
     $db_result = mysqli_query($db_link, $db_query);
 
-    $payout_total = 0;
+    $payout_total = (float)0;
 
     while ($row = mysqli_fetch_array($db_result)) {
-        $payout_total = $payout_total + $row['payout'];
+        $payout_total = $payout_total + (float)$row['payout'];
     }
 
-    return $payout_total;
+    return (float)$payout_total;
 }
